controllo sicurezza backend

+++ hackhub_backend/src/main/java/it/unicam/cs/hackhub/security/JwtUtil.java
@@ -52,7 +52,7 @@ public class JwtUtil {
     private String secret;
 
     @Value("${jwt.expiration}")
-    private int expiration;
+    private long expiration;
 
     private SecretKey key;
